requirements.txt
Flask==3.1.1
Flask-Limiter==3.12
Flask-Cors==5.0.1
PyJWT==2.10.1
supabase==2.15.1
python-dotenv==1.1.0
gunicorn==23.0.0
